Move entrypoints to individual folders and update config/entryPoints.js
