+++ src/utilities/subscriptions/urql/index.ts
@@ -1,13 +1,13 @@
-import { Observable } from "../index.js";
+import { Observable } from "../../index.js";
 import {
   handleError,
   readMultipartBody,
-} from "../../link/http/parseAndCheckHttpResponse.js";
-import { maybe } from "../index.js";
-import { serializeFetchParameter } from "../../core/index.js";
-import type { Body } from "../../link/http/selectHttpOptionsAndBody.js";
-import { generateOptionsForMultipartSubscription } from "./shared.js";
-import type { CreateMultipartSubscriptionOptions } from "./shared.js";
+} from "../../../link/http/parseAndCheckHttpResponse.js";
+import { maybe } from "../../index.js";
+import { serializeFetchParameter } from "../../../core/index.js";
+import type { Body } from "../../../link/http/selectHttpOptionsAndBody.js";
+import { generateOptionsForMultipartSubscription } from "../shared.js";
+import type { CreateMultipartSubscriptionOptions } from "../shared.js";
 
 const backupFetch = maybe(() => fetch);
